requirements.txt
amqp==5.0.9
asgiref==3.5.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.3
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
Django==4.0.2
idna==3.3
kombu==5.2.3
prompt-toolkit==3.0.28
python-decouple==3.6
pytz==2021.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
